+++ packages/transport/rn-hid/src/api/helpers/base64Utils.ts
@@ -1,14 +1,15 @@
+import {
+  base64StringToBuffer,
+  bufferToBase64String,
+} from "@ledgerhq/device-management-kit";
+
 /**
  * Encodes a Uint8Array to a Base64-encoded string.
  * @param byteArray - The Uint8Array to encode.
  * @returns A Base64-encoded string representing the byte array.
  */
 export function uint8ArrayToBase64(byteArray: Uint8Array): string {
-  const binary = byteArray.reduce((acc, byte) => {
-    acc += String.fromCharCode(byte);
-    return acc;
-  }, "");
-  return btoa(binary);
+  return bufferToBase64String(byteArray);
 }
 
 /**
@@ -17,10 +18,9 @@ export function uint8ArrayToBase64(byteArray: Uint8Array): string {
  * @returns A Uint8Array representing the decoded byte array.
  */
 export function base64ToUint8Array(base64: string): Uint8Array {
-  const binary = atob(base64);
-  const bytes = new Uint8Array(binary.length);
-  for (let i = 0; i < binary.length; i++) {
-    bytes[i] = binary.charCodeAt(i);
+  const bytes = base64StringToBuffer(base64);
+  if (bytes === null) {
+    throw new Error(`Invalid Base64 string: ${base64}`);
   }
   return bytes;
 }